Fix GH-16131: prevent subclass ctor/driver connection string mismatch
We now use the create driver-specific codepath without an object to
change, so we go through the exceptions thrown there when there's any
mismatches between the class being used and the connection string.

The class entry fetch in the entry points was also wrong and did not
get a valid class entry, this was also fixed.

A test was added for this, although I'm not pleased by the fact it needs
two real PDO drivers. A better way to test this would be nice, although
it does match the original sample case.